Remove GPL licensed code to get rid of the infection
The validity of this action depends on your definitions of modified and
new code.
I feel that I have changed serial.vhd and BRG.vhd so much that any
original content is basically non-existant, and as such, I get ownership
over that code, and can decide to license it under an alternative (less
infectious) license.

To help comply, I have removed the rest of the muart code which I have
not re-written. And I have left a note to mention the original author in
the 2 affected files, stating that the code is loosely based on his.

Please feel free to check the history of these 2 files and decide as to
whether you agree with me. But hey, I have just removed the GPL notices.
So hah.